// frontendapp/packages/lp-toolkit/src/lib/models/modal-footer-config.ts
export interface ModalFooterConfig {
  showConfirmButton?: boolean;
  confirmButtonText?: string;
  confirmButtonIcon?: string;
  confirmButtonDisabledState?: boolean;
  confirmButtonLoadingState?: boolean;
  showCancelButton?: boolean;
  cancelButtonText?: string;
  cancelButtonIcon?: string;
  cancelButtonDisabledState?: boolean;
  cancelButtonLoadingState?: boolean;
  showDeleteButton?: boolean;
  deleteButtonText?: string;
  deleteButtonIcon?: string;
  deleteButtonDisabledState?: boolean;
  deleteButtonLoadingState?: boolean;
}